+++ src/lazy.c
@@ -4,6 +4,7 @@
 SEXP promise_as_lazy(SEXP promise, SEXP env, int follow_symbols) {
   // recurse until we find the real promise, not a promise of a promise
   // never go past the global environment
+  int i = 0;
   while(TYPEOF(promise) == PROMSXP && env != R_GlobalEnv) {
     if (PRENV(promise) == R_NilValue) {
       Rf_error("Promise has already been forced");
@@ -15,12 +16,14 @@ SEXP promise_as_lazy(SEXP promise, SEXP env, int follow_symbols) {
     // If the promise is threaded through multiple functions, we'll
     // get some symbols along the way. If the symbol is bound to a promise
     // keep going on up
-    if (follow_symbols && TYPEOF(promise) == SYMSXP) {
+    // Unless we have reached follow_symbols loops, then we should return the symbol
+    if (TYPEOF(promise) == SYMSXP && (follow_symbols == NA_INTEGER || i < follow_symbols)) {
       SEXP obj = findVar(promise, env);
       if (TYPEOF(obj) == PROMSXP) {
         promise = obj;
       }
     }
+    i++;
   }
 
   // Make named list for output
@@ -42,14 +45,14 @@ SEXP promise_as_lazy(SEXP promise, SEXP env, int follow_symbols) {
 
 SEXP make_lazy(SEXP name, SEXP env, SEXP follow_symbols_) {
   SEXP promise = findVar(name, env);
-  int follow_symbols = asLogical(follow_symbols_);
+  int follow_symbols = asInteger(follow_symbols_);
 
   return promise_as_lazy(promise, env, follow_symbols);
 }
 
 SEXP make_lazy_dots(SEXP env, SEXP follow_symbols_) {
   SEXP dots = findVar(install("..."), env);
-  int follow_symbols = asLogical(follow_symbols_);
+  int follow_symbols = asInteger(follow_symbols_);
 
   // Figure out how many elements in dots
   int n = 0;