+++ NEWS.md
@@ -1,5 +1,8 @@
 # lazyeval 0.1.10.9000
 
+* Added the option to .follow_symbols an integer number of times in lazy 
+  (#39, @gergness)
+
 * Fixed an infinite loop in `lazy_dots(.follow_symbols = TRUE)` (#22, #24)
 
 * `lazy()` now fails with an informative error when it is applied on